Enable static site rendering with preview
Preview was added via native next.js support, it
just has to start through redirect over an api
call.

Not found page was also added, along with an
environment variable that is true when api key
is present. Just a heading was changed so far.

This is the first attempt. The code has to be
deployed in order to verify preview works as
intended.